refactor: resolve #756 - extract shared extractCodeBlocks test helper from tutorial lesson tests
Closes #756
