+++ test/ide/basicsLessons.test.ts
@@ -10,6 +10,8 @@ import { describe, expect, it } from 'vitest'
 import { basicsLessons } from '@/features/ide/tutorial/index'
 import type { Lesson } from '@/features/ide/tutorial/types'
 
+import { extractCodeBlocks } from '../helpers/extractCodeBlocks'
+
 // ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
@@ -26,22 +28,6 @@ const LESSON_KEYWORDS: Array<RegExp> = [
   /\bIF\b.*\bTHEN\b/,
   /\bFOR\b.*\bTO\b/,
 ]
-
-/**
- * Extracts all fenced code block contents from a markdown string.
- */
-function extractCodeBlocks(markdown: string): string[] {
-  const blocks: string[] = []
-  const regex = /```(?:\w*)\n([\s\S]*?)```/g
-  let match: RegExpExecArray | null
-  while ((match = regex.exec(markdown)) !== null) {
-    const code = match[1]?.trim()
-    if (code != null && code.length > 0) {
-      blocks.push(code)
-    }
-  }
-  return blocks
-}
 
 // ---------------------------------------------------------------------------
 // Tests
